requirements.txt
astroid==2.6.5
autoflake==1.4
black==21.7b0
email-validator==1.1.3
flask-babelex==0.9.4
flask-celery-helper==1.1.0
flask-cors==3.0.10
flask-jwt-extended==4.2.3
flask-migrate==3.0.1
flask-script==2.0.6
flask-user==1.0.2.2
isort==5.9.2
mccabe==0.6.1
mlflow==1.19.0
pip-chill==1.0.1
redis==3.5.3
toml==0.10.2
